backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.1.7
fastapi==0.115.5
h11==0.14.0
idna==3.2
Pillow==8.3.1
pydantic==2.10.2
pydantic_core==2.27.1
replicate==1.0.4
requests==2.26.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==1.26.6
uvicorn==0.32.1
